+++ image-viewer/extractors/fastpic.js
@@ -3,32 +3,32 @@ import { getURLFromPage } from './helpers'
 /** @typedef {import('../url-extractor').Extractor} Extractor */
 
 /*
-  link:       http://fastpic.ru/view/104/2018/0429/f1a539f2e9edd3e0d70cac3dcf316466.jpg.html
-  thumbnail:  http://i104.fastpic.ru/thumb/2018/0429/66/f1a539f2e9edd3e0d70cac3dcf316466.jpeg
-  image:      https://i104.fastpic.ru/big/2018/0429/66/f1a539f2e9edd3e0d70cac3dcf316466.jpg
+  link:       http://fastpic.org/view/104/2018/0429/f1a539f2e9edd3e0d70cac3dcf316466.jpg.html
+  thumbnail:  http://i104.fastpic.org/thumb/2018/0429/66/f1a539f2e9edd3e0d70cac3dcf316466.jpeg
+  image:      https://i104.fastpic.org/big/2018/0429/66/f1a539f2e9edd3e0d70cac3dcf316466.jpg
 */
 
 /** @type {Extractor} */
 export const fastpic = {
   name: 'FastPic',
-  linkRegExp: /^http.?:\/\/fastpic\.ru\/view/,
-  // TODO: There might be many matches (like https://fastpic.org/view/116/2022/0404/_203ed559db9fd49e5aa706cb9d1fae2a.jpg.html)
-  imageURLRegExp: /src="(?<url>[^"]+)" class="image img-fluid"/,
+  linkRegExp: /^http.?:\/\/fastpic\.(?:ru|org)\/view/,
+  imageURLRegExp: /src="(?<url>http[^"]+)" class="image img-fluid"/,
   getURL: getURLFromPage,
 }
 
 /*
-  link:       https://i104.fastpic.ru/big/2018/0429/66/f1a539f2e9edd3e0d70cac3dcf316466.jpg
-    i(104) big(2018/0429/) f1a539f2e9edd3e0d70cac3dcf316466.jpg
-  link:       http://fastpic.ru/view/ + 104 + /2018/0429/ + f1a539f2e9edd3e0d70cac3dcf316466.jpg + .html
+  link:       https://i104.fastpic.org/big/2018/0429/66/f1a539f2e9edd3e0d70cac3dcf316466.jpg
+                      i(104) big(2018/0429/) f1a539f2e9edd3e0d70cac3dcf316466.jpg
+  link:       http://fastpic.org/view/ + 104 + /2018/0429/ + f1a539f2e9edd3e0d70cac3dcf316466.jpg + .html
 */
 
-const URL_PARTS_REGEXP = /i(\d+).+big(\/\d+\/\d+\/).+\/([^/]+)$/
+/** Groups: index, top-level domain, data, filename. */
+const URL_PARTS_REGEXP = /i(\d+).+\.(ru|org)\/big(\/\d+\/\d+\/).+\/([^/]+)$/
 
 /** @type {Extractor} */
 export const fastpicDirect = {
   name: 'FastPic (direct link)',
-  linkRegExp: /fastpic\.ru\/big/,
+  linkRegExp: /fastpic\.(?:ru|org)\/big/,
 
   async getURL(link) {
     let hostLink = link.url
@@ -46,9 +46,10 @@ export const fastpicDirect = {
       }
     }
 
-    const [, index, date, filename] = URL_PARTS_REGEXP.exec(hostLink)
+    const [, index, domain, date, filename] =
+      URL_PARTS_REGEXP.exec(hostLink) || []
 
-    const url = `https://fastpic.ru/view/${index}${date}${filename}.html`
+    const url = `https://fastpic.${domain}/view/${index}${date}${filename}.html`
 
     return fastpic.getURL({ ...link, url }, fastpic)
   },